Create 2021-04-01_Florence_Nightingale_evening_program.md
Added an evening program, Florence Nightingale, to posts. Not sure how to get the image set in place without first adding new file, committing it and then going out to find the image name in assets, copying it and then going back in the new file, pasting it and committing again.